refactor: move scales and chords to menus and separate the two
