+++ menus/keys/scales.lua
@@ -701,218 +701,3 @@ function MinorPentatonic()
     hs.eventtap.keyStroke({}, "Up", 0);
     hs.eventtap.keyStroke({}, "Up", 0)
 end
-
--- CHORDS START HERE --
-
-function Octaves()
-    hs.eventtap.keyStroke({"cmd"}, "C", 0);
-    hs.eventtap.keyStroke({"cmd"}, "V", 0)
-    for i = 1, 12, 1 do
-        hs.eventtap.keyStroke({}, "Up", 0)
-    end
-end
-
-function Powerchord()
-    hs.eventtap.keyStroke({"cmd"}, "C", 0);
-    hs.eventtap.keyStroke({"cmd"}, "V", 0)
-    for i = 1, 7, 1 do
-        hs.eventtap.keyStroke({}, "Up", 0)
-    end
-end
-
-function Maj()
-    hs.eventtap.keyStroke({"cmd"}, "C", 0);
-    hs.eventtap.keyStroke({"cmd"}, "V", 0)
-    for i = 1, 4, 1 do
-        hs.eventtap.keyStroke({}, "Up", 0)
-    end
-    hs.eventtap.keyStroke({"cmd"}, "C", 0);
-    hs.eventtap.keyStroke({"cmd"}, "V", 0)
-    for i = 1, 3, 1 do
-        hs.eventtap.keyStroke({}, "Up", 0)
-    end
-end
-
-function Min()
-    hs.eventtap.keyStroke({"cmd"}, "C", 0);
-    hs.eventtap.keyStroke({"cmd"}, "V", 0)
-    for i = 1, 3, 1 do
-        hs.eventtap.keyStroke({}, "Up", 0)
-    end
-    hs.eventtap.keyStroke({"cmd"}, "C", 0);
-    hs.eventtap.keyStroke({"cmd"}, "V", 0)
-    for i = 1, 4, 1 do
-        hs.eventtap.keyStroke({}, "Up", 0)
-    end
-end
-
-function Aug()
-    hs.eventtap.keyStroke({"cmd"}, "C", 0);
-    hs.eventtap.keyStroke({"cmd"}, "V", 0)
-    for i = 1, 4, 1 do
-        hs.eventtap.keyStroke({}, "Up", 0)
-    end
-    hs.eventtap.keyStroke({"cmd"}, "C", 0);
-    hs.eventtap.keyStroke({"cmd"}, "V", 0)
-    for i = 1, 4, 1 do
-        hs.eventtap.keyStroke({}, "Up", 0)
-    end
-end
-
-function Dim()
-    hs.eventtap.keyStroke({"cmd"}, "C", 0);
-    hs.eventtap.keyStroke({"cmd"}, "V", 0)
-    for i = 1, 3, 1 do
-        hs.eventtap.keyStroke({}, "Up", 0)
-    end
-    hs.eventtap.keyStroke({"cmd"}, "C", 0);
-    hs.eventtap.keyStroke({"cmd"}, "V", 0)
-    for i = 1, 3, 1 do
-        hs.eventtap.keyStroke({}, "Up", 0)
-    end
-end
-
-function Maj7()
-    hs.eventtap.keyStroke({"cmd"}, "C", 0);
-    hs.eventtap.keyStroke({"cmd"}, "V", 0)
-    for i = 1, 4, 1 do
-        hs.eventtap.keyStroke({}, "Up", 0)
-    end
-    hs.eventtap.keyStroke({"cmd"}, "C", 0);
-    hs.eventtap.keyStroke({"cmd"}, "V", 0)
-    for i = 1, 3, 1 do
-        hs.eventtap.keyStroke({}, "Up", 0)
-    end
-    hs.eventtap.keyStroke({"cmd"}, "C", 0);
-    hs.eventtap.keyStroke({"cmd"}, "V", 0)
-    for i = 1, 4, 1 do
-        hs.eventtap.keyStroke({}, "Up", 0)
-    end
-end
-
-function Min7()
-    hs.eventtap.keyStroke({"cmd"}, "C", 0);
-    hs.eventtap.keyStroke({"cmd"}, "V", 0)
-    for i = 1, 3, 1 do
-        hs.eventtap.keyStroke({}, "Up", 0)
-    end
-    hs.eventtap.keyStroke({"cmd"}, "C", 0);
-    hs.eventtap.keyStroke({"cmd"}, "V", 0)
-    for i = 1, 4, 1 do
-        hs.eventtap.keyStroke({}, "Up", 0)
-    end
-    hs.eventtap.keyStroke({"cmd"}, "C", 0);
-    hs.eventtap.keyStroke({"cmd"}, "V", 0)
-    for i = 1, 3, 1 do
-        hs.eventtap.keyStroke({}, "Up", 0)
-    end
-end
-
-function Dom7()
-    hs.eventtap.keyStroke({"cmd"}, "C", 0);
-    hs.eventtap.keyStroke({"cmd"}, "V", 0)
-    for i = 1, 4, 1 do
-        hs.eventtap.keyStroke({}, "Up", 0)
-    end
-    hs.eventtap.keyStroke({"cmd"}, "C", 0);
-    hs.eventtap.keyStroke({"cmd"}, "V", 0)
-    for i = 1, 3, 1 do
-        hs.eventtap.keyStroke({}, "Up", 0)
-    end
-    hs.eventtap.keyStroke({"cmd"}, "C", 0);
-    hs.eventtap.keyStroke({"cmd"}, "V", 0)
-    for i = 1, 3, 1 do
-        hs.eventtap.keyStroke({}, "Up", 0)
-    end
-end
-
-function Maj9()
-    hs.eventtap.keyStroke({"cmd"}, "C", 0);
-    hs.eventtap.keyStroke({"cmd"}, "V", 0)
-    for i = 1, 4, 1 do
-        hs.eventtap.keyStroke({}, "Up", 0)
-    end
-    hs.eventtap.keyStroke({"cmd"}, "C", 0);
-    hs.eventtap.keyStroke({"cmd"}, "V", 0)
-    for i = 1, 3, 1 do
-        hs.eventtap.keyStroke({}, "Up", 0)
-    end
-    hs.eventtap.keyStroke({"cmd"}, "C", 0);
-    hs.eventtap.keyStroke({"cmd"}, "V", 0)
-    for i = 1, 4, 1 do
-        hs.eventtap.keyStroke({}, "Up", 0)
-    end
-    hs.eventtap.keyStroke({"cmd"}, "C", 0);
-    hs.eventtap.keyStroke({"cmd"}, "V", 0)
-    for i = 1, 3, 1 do
-        hs.eventtap.keyStroke({}, "Up", 0)
-    end
-end
-
-function Min9()
-    hs.eventtap.keyStroke({"cmd"}, "C", 0);
-    hs.eventtap.keyStroke({"cmd"}, "V", 0)
-    for i = 1, 3, 1 do
-        hs.eventtap.keyStroke({}, "Up", 0)
-    end
-    hs.eventtap.keyStroke({"cmd"}, "C", 0);
-    hs.eventtap.keyStroke({"cmd"}, "V", 0)
-    for i = 1, 4, 1 do
-        hs.eventtap.keyStroke({}, "Up", 0)
-    end
-    hs.eventtap.keyStroke({"cmd"}, "C", 0);
-    hs.eventtap.keyStroke({"cmd"}, "V", 0)
-    for i = 1, 3, 1 do
-        hs.eventtap.keyStroke({}, "Up", 0)
-    end
-    hs.eventtap.keyStroke({"cmd"}, "C", 0);
-    hs.eventtap.keyStroke({"cmd"}, "V", 0)
-    for i = 1, 4, 1 do
-        hs.eventtap.keyStroke({}, "Up", 0)
-    end
-end
-
-function Fold3()
-    hs.eventtap.keyStroke({"cmd"}, "C", 0);
-    hs.eventtap.keyStroke({"cmd"}, "V", 0);
-    hs.eventtap.keyStroke({}, "Up", 0);
-    hs.eventtap.keyStroke({}, "Up", 0)
-    hs.eventtap.keyStroke({"cmd"}, "C", 0);
-    hs.eventtap.keyStroke({"cmd"}, "V", 0);
-    hs.eventtap.keyStroke({}, "Up", 0);
-    hs.eventtap.keyStroke({}, "Up", 0)
-end
-
-function Fold7()
-    hs.eventtap.keyStroke({"cmd"}, "C", 0);
-    hs.eventtap.keyStroke({"cmd"}, "V", 0);
-    hs.eventtap.keyStroke({}, "Up", 0);
-    hs.eventtap.keyStroke({}, "Up", 0)
-    hs.eventtap.keyStroke({"cmd"}, "C", 0);
-    hs.eventtap.keyStroke({"cmd"}, "V", 0);
-    hs.eventtap.keyStroke({}, "Up", 0);
-    hs.eventtap.keyStroke({}, "Up", 0)
-    hs.eventtap.keyStroke({"cmd"}, "C", 0);
-    hs.eventtap.keyStroke({"cmd"}, "V", 0);
-    hs.eventtap.keyStroke({}, "Up", 0);
-    hs.eventtap.keyStroke({}, "Up", 0)
-end
-
-function Fold9()
-    hs.eventtap.keyStroke({"cmd"}, "C", 0);
-    hs.eventtap.keyStroke({"cmd"}, "V", 0);
-    hs.eventtap.keyStroke({}, "Up", 0);
-    hs.eventtap.keyStroke({}, "Up", 0)
-    hs.eventtap.keyStroke({"cmd"}, "C", 0);
-    hs.eventtap.keyStroke({"cmd"}, "V", 0);
-    hs.eventtap.keyStroke({}, "Up", 0);
-    hs.eventtap.keyStroke({}, "Up", 0)
-    hs.eventtap.keyStroke({"cmd"}, "C", 0);
-    hs.eventtap.keyStroke({"cmd"}, "V", 0);
-    hs.eventtap.keyStroke({}, "Up", 0);
-    hs.eventtap.keyStroke({}, "Up", 0)
-    hs.eventtap.keyStroke({"cmd"}, "C", 0);
-    hs.eventtap.keyStroke({"cmd"}, "V", 0);
-    hs.eventtap.keyStroke({}, "Up", 0);
-    hs.eventtap.keyStroke({}, "Up", 0)
-end
